courses/graphacademy/requirements.txt
langchain==0.1.10
langchain-openai
langchainhub
openai==1.13.3
python-dotenv
youtube-search
neo4j
streamlit
